mcp_a2a_master/pyproject.toml
[project]
name = "mcp-a2a-master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.25",
    "google-adk>=1.27.2",
    "mcp[cli]>=1.26.0",
]
